Added an API
I added a superuser so I could add data from the Admin page.  I create one field as a test for the API.  I created the app api url and linked it the project urls.  Then, I created the serializer so one can do querysets on the data.  ex: ?id=1
Then, I created the json views so the API knew what data to display on the screen.