Adjusted database/models.
I changed the model so that it would hold a first name and a last name field.  Then, I had to change the serializers and json_views so that it would match the model.  I inputted some data as an admin and viewed it as .json data
ex: /api/v1/berlin?format=json
I copied all of the data and dumped it into an initial_data.json so I would have beginning data just in case something happened to my models or database.  Pushing so I don't mess anything up.