Big commit for the StudyAbroad project.
Templates- Created an aboutus, main, and profiles html templates.  I will later go back and make them better.  For now, if you click on the lik on the nav bar, it will navigate to these pages.
Admin- Adjusted the admin so that I can see the Latitude and Longitude in the map model.
Json_views & serializers-  Made these both geo friendly for the map model.
Base-  I used the bootstrap to make the nav bar.  I will need to format it to go with my project.  It is still a work in progress.

Still a work in progress, but I fixed the database and rigged the admin to show my lat & long.