Urls update and map model fixed.
I changed the url so I have a plain main page-- I can just type in 127.0.0.800/ and it'll take me to the main page.
I also changed the map model to work the way that I want it to.  The admin can go into the model and place any point on the map; however, they can override this point by typing in the Latitude and Longitude that they want.  I originally had:
        self.Latitude = self.geom.y
        self.Longitude = self.geom.x

        but it need to be this:

         self.geom.y = self.Latitude
        self.geom.x = self.Longitude

  Next step in the backend: work on adding pictures