Updated API and added student profiles
I changed the serializers and json_views so that the API will show all of the data associated with the student model.  I also added a student_detail.html... When the user goes to browse profiles, a list template will display all of the students in the database.  I want it so that when the user clicks on a name, it will go to that profile.  Right now it does that, but it is not showing any data for that student.

I also changed the main page to just be the "127.0.0.1:8000" instead of "127.0.0.1:8000/berlin/main" --therefore I had to change all of the links in the navbar and carousel buttons