Added map and changed navbar.css
I created a map in mapbox and linked all of the correct sites to the base.html
popular_places.html is where you can find the map displayed. The map currently goes straight to the berlin bookmark.

Detailed student-  I added a url called "profiles" where it displays all of the students in one page.  The detailed view then goes to the student for each id.  The url is then changed to : http://127.0.0.1:8000/profiles/1/
where "1" is the pk for the student.  This detailed view displays the name, major, school, and classes that were taken in the class.

Navbar-  I added the static folder and added a navbar.css to add color to my navbar.  The colors are maroon, white, and black.  I added a link to it on my base.html

Main page- I added three circle images to the jumbotron.  I plan on adding pictures of Berlin on the front page-- I will eventually make these clickable and take the user to another page.  I also added a form to the front page.  It doesn't work yet, but this is where the user will sign in to add photos or popular places on the map.