data-  This data folder holds the shapefile of the colleges that the students go to that went to Berlin
templates-  This is just a simple html file that loads in the mapbox view of my Georgia bookmark.  I plan on using the API to pull in the points of the colleges that I geocoded.  I want to map out and see what school has the most students that went to Berlin

admin- I simply just added the Georgia model to see it in the admin page.  I might take it out since I do not plan on changing or adding any data

api_urls, json_views, serializers-  After I loaded in the shapefile,  I created an api url to use to see the data of the shapefile.  I also created the correct json_view and serializer for it.  To access the API, use the url : /api/v1/georgia

load-  I used the GeoDjango tutorial and walked through how to add a shapefile into my database.
https://docs.djangoproject.com/en/dev/ref/contrib/gis/tutorial/

base.html nav bar -- I didn't add that much.  I changed the "stuff" place holder and added "Georgia maps" to the nav bar.  When clicked, it will display the georgia.html page which shows the mapbox georgia map that I created. This is where people will view the shapefile that I loaded.