Added Registration and changed navbar colors
I changed the navbar colors so they are a nice grassy green instead of a maroon red.  The nav bar now matches the "sign in" button on the main page.  On the main page, I temporally took out the pictures on the front.  I will eventually go in and add photos of my own.  I used the Django-Registration-redux for my registration.  I made sure it was installed in my settings, and added it to my settings.py.  Next, I followed the directions on the documentation page and added the login.html, logout.html, registration_complete.html, and registration_form.html.
I also added the "login" and "register" buttons to the nav bar.  If the user is already logged in, it will say logout instead.  If the user is the admin of the site, a nice "admin" button will appear.  I wanted there to be a form on the main page where the user can just input from there, but I couldn't get that to work so I just added a button.  In addition, I added "LOGIN_REDIRECT_URL = '/'"  in my settings so once the user logs in, it redirects them to the main page.
Next, I will need to make the browse profiles page only accessible to those who have an account.  Still a work in progress, but I still have a lot to work.