Displaying data on Georgia Map
I added some javascript/jQuery for my Georgia map.  When the user goes to the Georgia map, the map of Georgia is displayed with all of the schools on the map.  I made the schools red circle to make them stand out.  I had the default pointer, but wanted to change it up a little.  When the user clicks on the school, a little pop up appears and it says how many number of students from that school went to Berlin.  I would like for the link that pops up to display all of the students from that school that went or the school's website.  I'm still undecided on that part.  This spatial model is pretty much done.  Now that I have tested out adding and loading shapefiles and displaying data on the map from the API, I will compile all of the popular_places and do it again but for that spatial model.