Popular Places map works
I changed the name of the popular places map to "popular_places_map" so I could use the name "popular_places" to list out all of the places in my database.  When you click on the name of the place, it will take you to the detailed view of the place with the name, geom, lat, and long.  I think I will need to destroy my database so I can add a description field and a photo field to display on the detailed view.  So since I changed the name, I had to change the urls and views names.  I also added the detailed view to the urls and views-- I pretty much just copied what I had for the students detailed view and changed the names.  Since the map is now displaying on "popular_places_map", I went into the base.html and changed the path for the popular places map.

While on the map of the popular places, the places are orange circles.  When you click on the point, the name pops up.  If you click on the name, it will re-direct you to the detailed view.

Since I am thinking about blowing up my database next, I copied the student data and placed it in the initial_data.json so I can just re-add it later.