Adding pictures to the popular_places detail page
I did a lot of research in how to display pictures on the page without using Amazon S3.  I created an Image model and put it in the admin.  When the user goes to the admin, they click on the place they want to add the image to.  The image will then save with that foreign key.  I have the popular_places detail page pull in the images with its name.  The images are being saved in Pycharm but not in the database...?  I'm a little confused on that part, but I plan on this just being a project and not deploying it.  I also adjusted the urls and media roots so that they fit with the images.  I decided to add a type and description to the popular places model; I'm not sure why I didn't do that early.  Now, I will re-do my database and reload the data.  Then I can upload some pictures to go along with the pages.  After that, I will re-design the popular_pages_detail page, and then try and maybe work with the map.  I'm definitely on the ball--looking forward to the finished project.