Took away features to make people create an account
I adjusted my site so that people will have to create an account and log in to access the full site.  When a user logs in, they will have access to the full navigation bar (Home, About Us, Browse Profiles, and Maps).  When they go to the home page, they will be greeted with a welcome message and a button that links them to the Map of Berlin.  If a user is not logged in, they will only have "Home and About Us" on their nav bar.  On the Home page, they will be greeted with a brief message about what the site is about and a button that will forward them to a registration page.