Create form for Popular_Places and html clean up
I created a form for my popular places model.  When the user goes to the popular places map, an orange button will appear at the top of the screen that says "Add a Point."  When the user clicks this, it takes them to "/create/" which has the form for them to fill out.  Right now, I have it displaying correctly, and if the user does input something, it gives them a successful url (just the home page); however, it isn't saving to the database...I will need to look at it next.

In my Main.html, I fixed how it looks depending who is looking at it.  When a user isn't logged in, a green "Sign In" button appears.  When the user logs in, that button disappears.  I also made the popular places detail paged center.  As far as database goes, I re-entered the student data and added some pictures to my local machine to display on the popular_places detail page.